+++ product-service/.gitignore
@@ -8,4 +8,4 @@ jspm_packages
 # esbuild directories
 .esbuild
 
-.swagger
+/swagger
