+++ product-service/src/libs/load-products.ts
