--- product-service/swagger/swagger-json.d.ts
@@ -1,1 +1,0 @@
-declare const swagger: any;
