--- product-service/swagger/swagger-json.js
@@ -1,9 +1,0 @@
-"use strict";
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const swagger = require('./swagger');
-exports.handler = async () => {
-    return {
-        statusCode: 200,
-        body: JSON.stringify(swagger),
-    };
-};
