--- product-service/swagger/swagger.js
@@ -1,107 +1,0 @@
-// this file was generated by serverless-auto-swagger
-            module.exports = {
-  "swagger": "2.0",
-  "info": {
-    "title": "product-service",
-    "version": "1"
-  },
-  "paths": {
-    "/products": {
-      "get": {
-        "summary": "getProductsList",
-        "description": "",
-        "operationId": "getProductsList.get.products",
-        "consumes": [
-          "application/json"
-        ],
-        "produces": [
-          "application/json"
-        ],
-        "parameters": [],
-        "responses": {
-          "200": {
-            "description": "Successful API response",
-            "schema": {
-              "$ref": "#/definitions/ProductList"
-            }
-          }
-        }
-      }
-    },
-    "/products/{id}": {
-      "get": {
-        "summary": "getProductById",
-        "description": "",
-        "operationId": "getProductById.get.products/{id}",
-        "consumes": [
-          "application/json"
-        ],
-        "produces": [
-          "application/json"
-        ],
-        "parameters": [
-          {
-            "name": "id",
-            "in": "path",
-            "required": true,
-            "type": "string"
-          }
-        ],
-        "responses": {
-          "200": {
-            "description": "Successful API response",
-            "schema": {
-              "$ref": "#/definitions/Product"
-            }
-          }
-        }
-      }
-    }
-  },
-  "definitions": {
-    "Product": {
-      "properties": {
-        "count": {
-          "title": "Product.count",
-          "type": "number"
-        },
-        "description": {
-          "title": "Product.description",
-          "type": "string"
-        },
-        "id": {
-          "title": "Product.id",
-          "type": "string"
-        },
-        "price": {
-          "title": "Product.price",
-          "type": "number"
-        },
-        "title": {
-          "title": "Product.title",
-          "type": "string"
-        }
-      },
-      "required": [
-        "count",
-        "description",
-        "id",
-        "price",
-        "title"
-      ],
-      "additionalProperties": false,
-      "title": "Product",
-      "type": "object"
-    },
-    "ProductList": {
-      "items": {
-        "$ref": "#/definitions/Product",
-        "title": "ProductList.[]"
-      },
-      "title": "ProductList.[]",
-      "type": "array"
-    }
-  },
-  "securityDefinitions": {},
-  "host": "afgfgse3n7.execute-api.us-east-1.amazonaws.com/dev"
-};
